@@ -3,12 +3,13 @@
 
 """
 
-Maths
+Mathematics
 
     f_lin, f_lin0, f_lin1, gaussian, gaussian2D,
     rms, nanrms, std, nanstd, nanavg,
+    pix2sr, sr2arcsec2, rad2arcsec, hour2deg, deg2hour,
     icorr2ij, ij2icorr,
-    bsplinterpol
+    bsplinterp
 
 """
 
@@ -18,8 +19,11 @@ import scipy.interpolate as interpolate
 import warnings
 
 ## Local
-from utilities import InputError
+import utbox as UT
 
+
+## Python string formatting
+## https://pyformat.info
 
 def f_lin(x, A, B):
     '''
@@ -165,75 +169,151 @@ def nanavg(a, axis=None, weights=None, MaskedValue=np.nan):
 
     return avg
 
-def icorr2ij(Npar, Ncorr=None, upper=True):
+def pix2sr(X, CDELT):
+    '''
+    X pixel = Y sr
+    ------ INPUT ------
+    X                   float or ndarray in pixel
+    CDELT               float or ndarray (same dim if X is also ndarray)
+    '''
+    PFOV = abs(CDELT)
+    
+    return X * (PFOV * 2. * math.pi / 360.)**2.
+
+def sr2arcsec2(X):
+    '''
+    X sr = Y arcsec^2
+    '''
+    return X * (360. * 3600. / (2. * math.pi))**2.
+
+def rad2arcsec(X):
+    '''
+    X rad = Y arcsec
+    '''
+    return X * 360. * 3600. / (2. * math.pi)
+
+def hour2deg(h, m, s, deg, arcmin, arcsec):
+    '''
+    Hour angle to degree conversion
+    '''
+    ra = (h + m/60. + s/3600.) * 360./24.
+    if deg<0:
+        dec = -(-deg + arcmin/60. + arcsec/3600.)
+    else:
+        dec = deg + arcmin/60. + arcsec/3600.
+
+    return ra, dec
+
+def deg2hour(ra, dec):
+    '''
+    Degree to hour angle conversion
+    '''
+    h = math.floor(ra * 24./360.)
+    m = math.floor((ra*24./360. - h) * 60.)
+    s = ((ra*24./360. - h)*60 - m) * 60.
+    if dec<0:
+        deg = math.ceil(dec) # negtive
+        arcmin = -math.ceil((dec - deg) * 60.) # positive
+        arcsec = -((dec - deg)*60. + arcmin) * 60. # positive
+    else:
+        deg = math.floor(dec)
+        arcmin = math.floor((dec - deg) * 60.)
+        arcsec = ((dec - deg)*60. - arcmin) * 60.
+        
+    # print('{:d}h{:d}m{:04.2f}s, {:d}d{:d}m{:04.2f}s'.format(h,m,s,deg,arcmin,arcsec))
+
+    return h, m, s, deg, arcmin, arcsec
+
+def icorr2ij(Npar, icorr=None, upper=True):
     '''
     Get the pair of parameter indices and their correlation coefficient index
     
     ------ INPUT ------
-    Npar                Number of parameters
-    Ncorr               Number of correlation coefficients (Default: None)
+    Npar                Number of correlated parameters
+    icorr               Index of correlation coefficients (Default: None)
+                          None - return all ij pairs
     upper               Upper or lower triangles (Default: True)
     ------ OUTPUT ------
-    ij                  Indices with a dimension of (Ncorr,2)
+    ij                  Index pair
     '''
-    if Ncorr is None:
-        # Ncorr = math.comb(Npar,2) # Python v3.8+
-        Ncorr = int( Npar*(Npar-1)/2 )
+    Ncorr = math.comb(Npar,2) # Python >=3.8
+    # Ncorr = int( Npar*(Npar-1)/2 ) # alternative
     ij = np.empty((Ncorr,2), dtype=int)
-    icorr = 0
+    ic = 0
     for i in range(Npar):
         for j in range(Npar):
-            if upper:
-                if j>i:
-                    ij[icorr,:] = [i+1,j+1]
-                    icorr += 1
-            else:
-                if i>j:
-                    ij[icorr,:] = [j+1,i+1]
-                    icorr += 1
+            if j>i:
+                if upper:
+                    ij[ic,:] = [i+1,j+1]
+                else:
+                    ij[ic,:] = [j+1,i+1]
+                ic += 1
 
-    return ij
+    if icorr is None:
+        warnings.warn('<maths.icorr2ij> icorr is None! All icorr2ij pairs are returned. ')
+        return ij
+    else:
+        return ij[icorr-1]
 
-def ij2icorr(i, j, Npar, verbose=False):
+def ij2icorr(Npar, i, j, verbose=False):
     '''
     Get the correlation coefficient index given the pair of parameter indices
     
     ------ INPUT ------
+    Npar                Number of correlated parameters
     i                   Index of the first parameter (>= 1)
     j                   Index of the second parameter (>= 1)
-    Npar                Number of parameters
     verbose             Default: False
     ------ OUTPUT ------
     icorr               Indices of correlation coefficient
     '''
-    ## (i-1)*N-(i+1)C2+j
-    try:
-        icorr = 0
-        for ii in range(Npar):
-            for jj in range(Npar):
-                if j>i:
-                    if verbose:
-                        print('Upper triangle detected.')
-                    if jj>ii:
-                        icorr += 1
-                    if ii==i-1 and jj==j-1:
-                        raise StopIteration
-                elif i>j:
-                    if verbose:
-                        print('Lower triangle detected.')
-                    if ii>jj:
-                        icorr += 1
-                    if ii==i-1 and jj==j-1:
-                        raise StopIteration
-                else:
-                    warnings.warn('<maths.ij2icorr> i equals to j! icorr = 0')
-                    raise StopIteration
-
-    except StopIteration:
-        
+    icorr = np.empty((Npar,Npar), dtype=int)
+    ic = 0
+    for ii in range(Npar):
+        for jj in range(Npar):
+            if jj>ii:
+                ic += 1
+                icorr[jj,ii] = ic
+                icorr[ii,jj] = ic
+            elif ii==jj:
+                icorr[ii,jj] = 0
+                ## Return all ij2icorr pairs
+    if verbose:
+        if i<j:
+            print('Upper triangle detected.')
+        elif i>j:
+            print('Lower triangle detected.')
+        else:
+            warnings.warn('<maths.ij2icorr> i equals to j! All ij2icorr pairs are returned. ')
+    if i==j:
         return icorr
+    else:
+        return icorr[j-1,i-1]
 
-def bsplinterpol(x, y, x0):
+    ## Opt.1
+    # try:
+    #     icorr = 0
+    #     for ii in range(Npar):
+    #         for jj in range(Npar):
+    #             if j>i:
+    #                 if jj>ii:
+    #                     icorr += 1
+    #                 if ii==i-1 and jj==j-1:
+    #                     raise StopIteration
+    #             elif i>j:
+    #                 if ii>jj:
+    #                     icorr += 1
+    #                 if ii==i-1 and jj==j-1:
+    #                     raise StopIteration
+    #             else:
+    #                 raise StopIteration
+    # except StopIteration:
+    #     return icorr
+
+    ## Opt.2
+    # return (i-1)*Npar - math.comb(i+1,2) + j
+
+def bsplinterp(x, y, x0):
     '''
     Monte-Carlo propagated error calculator
     
@@ -261,11 +341,3 @@ def bsplinterpol(x, y, x0):
     bspl = interpolate.BSpline(t, c, k, extrapolate=False)
     
     return bspl(x0)
-
-"""
------------------------------- MAIN (test) ------------------------------
-"""
-if __name__ == "__main__":
-
-    pass
-
